requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
SQLAlchemy==2.0.35
PyYAML==6.0.2
numpy==1.26.4
opencv-python==4.10.0.84
